Fix tracing issue
