+++ benchmarks/benchmark_latency.py
@@ -13,10 +13,46 @@ from tqdm import tqdm
 from vllm import LLM, SamplingParams
 from vllm.inputs import PromptStrictInputs
 from vllm.model_executor.layers.quantization import QUANTIZATION_METHODS
-from vllm.utils import rpd_profiler_context, torch_profiler_context 
+from vllm.utils import rpd_trace
+from contextlib import contextmanager, nullcontext
 
 def main(args: argparse.Namespace):
     print(args)
+    
+    @contextmanager
+    def rpd_profiler_context(profile_dir: Optional[str] = None, trace_file_name = None):
+        trace_file_path = os.path.join(profile_dir, f"{trace_file_name}.rpd")
+        with rpd_trace(filename = f"{trace_file_path}", name = "run_to_completion", nvtx = True) as p:
+            yield p
+        p.rpd.top_totals()
+
+    @contextmanager
+    def torch_profiler_context(profile_dir: Optional[str] = None, trace_file_name = None):
+        p = torch.profiler.profile(
+                    activities=[
+                        torch.profiler.ProfilerActivity.CPU,
+                        torch.profiler.ProfilerActivity.CUDA,
+                    ],
+                    on_trace_ready=torch.profiler.tensorboard_trace_handler(
+                        str(profile_dir)))
+        p.start()
+        try:
+            with torch.no_grad():
+                yield p
+        finally:
+            p.stop()
+            print(p.key_averages().table(sort_by="self_cuda_time_total", row_limit=-1))
+
+    def get_profiling_context(profile_dir: Optional[str] = None, trace_file_name = None):
+         if args.profile_torch:
+             print(f"trace_file_name: {trace_file_name}")
+             return torch_profiler_context(profile_dir, trace_file_name)
+         elif args.profile_rpd:
+             print(f"trace_file_name: {trace_file_name}")
+             return rpd_profiler_context(profile_dir, trace_file_name)
+         else:
+             return nullcontext()
+
 
     # NOTE(woosuk): If the request cannot be processed in a single batch,
     # the engine will automatically process the request in multiple batches.
@@ -57,14 +93,6 @@ def main(args: argparse.Namespace):
     dummy_inputs: List[PromptStrictInputs] = [{
         "prompt_token_ids": batch
     } for batch in dummy_prompt_token_ids.tolist()]
-    
-    def get_profiling_context(profile_dir: Optional[str] = None, trace_file_name = None):
-         if args.profile_torch:
-             return torch_profiler_context(profile_dir, trace_file_name)
-         elif args.profile_rpd:
-             return rpd_profiler_context(profile_dir, trace_file_name)
-         else:
-             return nullcontext()
     
     def run_to_completion(profile_dir: Optional[str] = None, profiling_mode = None):
         if profile_dir:
