+++ vllm/utils.py
@@ -38,30 +38,6 @@ STR_DTYPE_TO_TORCH_DTYPE = {
     "fp8_e5m2": torch.uint8,
 }
 
-@contextmanager
-def rpd_profiler_context(profile_dir: Optional[str] = None, trace_file_name = None):
-    trace_file_path = os.path.join(profile_dir, f"{trace_file_name}.rpd")
-    with rpd_trace(filename = f"{trace_file_path}", name = "run_to_completion", nvtx = True) as p:
-        yield p
-    p.rpd.top_totals()
-
-@contextmanager
-def torch_profiler_context(profile_dir: Optional[str] = None, trace_file_name = None):
-    p = torch.profiler.profile(
-                activities=[
-                    torch.profiler.ProfilerActivity.CPU,
-                    torch.profiler.ProfilerActivity.CUDA,
-                ],
-                on_trace_ready=torch.profiler.tensorboard_trace_handler(
-                    str(profile_dir)))
-    p.start()
-    try:
-        with torch.no_grad():
-            yield p
-    finally:
-        p.stop()
-        print(p.key_averages().table(sort_by="self_cuda_time_total", row_limit=-1))
-
 class rpd_trace():
 
     def __init__(self, filename=None, name=None, nvtx=False, args=None, skip=False):
